Detect the ACM and MCE namespaces in case the default isn't used (#19)
The inspector was using hard coded namespaces for ACM and MCE.  This
caused problems when I tried to investigate a performance problem and
wanted to use the inspector. It did not recognize ACM in the `ocm`
namespace.  Please consider cases where I may have added the
MCE namespace or could have changed the pattern in your review.

Refs:
 - https://issues.redhat.com/browse/ACM-16868

Signed-off-by: Gus Parvin <gparvin@redhat.com>